pyproject.toml
[project]
name = "documentexplorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "gensim>=4.4.0",
    "markdown>=3.10",
    "pip>=25.3",
    "python-decouple>=3.8",
    "rake-nltk>=1.0.6",
    "requests>=2.32.5",
    "requests-cache>=1.2.1",
    "spacy>=3.8.11",
    "tenacity>=9.1.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "vadersentiment>=3.3.2",
]

[tool.uv.sources]
torch = { index = "pytorch_cpu" }

[[tool.uv.index]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
